refactor(recipe-view): Extract dates to component.
